+++ ReviewMonitor/AppDelegate.swift
@@ -19,16 +19,41 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     var blurEffectView: UIVisualEffectView?
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
+        window = UIWindow(frame: UIScreen.main.bounds)
+
         #if DEBUG
         #else
             Fabric.with([Answers.self, Crashlytics.self])
         #endif
 
-        window = UIWindow(frame: UIScreen.main.bounds)
         if AccountManger.getAccountArray().count == 0 {
             window?.rootViewController = LoginViewController(nibName: "LoginViewController", bundle: nil)
         } else {
             window?.rootViewController = UIStoryboard(name: "Main", bundle: nil).instantiateInitialViewController()
+        }
+        if UserDefaults.standard.bool(forKey: "useBiometrics") {
+            let context = LAContext()
+            var error: NSError?
+            addBlurView()
+            if context.canEvaluatePolicy(.deviceOwnerAuthenticationWithBiometrics, error: &error) {
+                let reason = "Please authenticate."
+                context.localizedFallbackTitle = ""
+                context.evaluatePolicy(.deviceOwnerAuthenticationWithBiometrics, localizedReason: reason) {
+                    [unowned self] success, authenticationError in
+
+                    DispatchQueue.main.async {
+                        if success {
+                            self.removeBlurView()
+                            print("Success")
+                        } else {
+                            print("FAIL")
+                        }
+                    }
+                }
+            } else {
+                removeBlurView()
+                print("NO Biometrics")
+            }
         }
         window?.makeKeyAndVisible()
         return true
@@ -40,59 +65,61 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         window?.makeKeyAndVisible()
     }
 
-    func addBlueView() {
+    func addBlurView() {
         let blurEffect = UIBlurEffect(style: UIBlurEffectStyle.dark)
         blurEffectView = UIVisualEffectView(effect: blurEffect)
         let view = window?.rootViewController?.view
         blurEffectView?.frame = (view?.bounds)!
         blurEffectView?.autoresizingMask = [.flexibleWidth, .flexibleHeight] // for supporting device rotation
-        view?.addSubview(blurEffectView!)
+        if !view!.subviews.contains(blurEffectView!) {
+            view?.addSubview(blurEffectView!)
+        }
     }
 
     func removeBlurView() {
         blurEffectView?.removeFromSuperview()
     }
 
-    //    func applicationWillEnterForeground(_ application: UIApplication) {
-    //        let touchIDManager = AccountManger()
-    //
-    //        addBlueView()
-    //
-    //        touchIDManager.authenticateUser(success: { () -> Void in
-    //            OperationQueue.main.addOperation({ () -> Void in
-    //                //                self.loadDada()
-    //                self.removeBlurView()
-    //            })
-    //        }, failure: { (evaluationError: NSError) -> Void in
-    //            switch evaluationError.code {
-    //            case LAError.Code.systemCancel.rawValue:
-    //                print("Authentication cancelled by the system")
-    //                //                self.statusLabel.text = "Authentication cancelled by the system"
-    //            case LAError.Code.userCancel.rawValue:
-    //                print("Authentication cancelled by the user")
-    //                //                self.statusLabel.text = "Authentication cancelled by the user"
-    //            case LAError.Code.userFallback.rawValue:
-    //                print("User wants to use a password")
-    //                //                self.statusLabel.text = "User wants to use a password"
-    //                // We show the alert view in the main thread (always update the UI in the main thread)
-    //                OperationQueue.main.addOperation({ () -> Void in
-    //                    //                    self.showPasswordAlert()
-    //                })
-    //            case LAError.Code.touchIDNotEnrolled.rawValue:
-    //                print("TouchID not enrolled")
-    //                //                self.statusLabel.text = "TouchID not enrolled"
-    //            case LAError.Code.passcodeNotSet.rawValue:
-    //                print("Passcode not set")
-    //                //                self.statusLabel.text = "Passcode not set"
-    //            default:
-    //                print("Authentication failed")
-    //                //                self.statusLabel.text = "Authentication failed"
-    //                OperationQueue.main.addOperation({ () -> Void in
-    //                    //                    self.showPasswordAlert()
-    //                })
-    //            }
-    //        })
-    //    }
+    var startBack = Date()
+    func applicationDidEnterBackground(_ application: UIApplication) {
+        print(Date())
+        startBack = Date()
+        addBlurView()
+    }
+
+    func applicationWillEnterForeground(_ application: UIApplication) {
+        if Date().timeIntervalSince(startBack) >= 30 && UserDefaults.standard.bool(forKey: "useBiometrics") {
+            print("Auth")
+            let context = LAContext()
+            var error: NSError?
+
+            if context.canEvaluatePolicy(.deviceOwnerAuthenticationWithBiometrics, error: &error) {
+                let reason = "Please authenticate."
+                context.localizedFallbackTitle = ""
+                context.evaluatePolicy(.deviceOwnerAuthenticationWithBiometrics, localizedReason: reason) {
+                    [unowned self] success, authenticationError in
+
+                    DispatchQueue.main.async {
+                        if success {
+                            self.removeBlurView()
+                            print("Success")
+                        } else {
+                            print("FAIL")
+                            //                            let ac = UIAlertController(title: "Authentication failed", message: "Sorry!", preferredStyle: .alert)
+                            //                            ac.addAction(UIAlertAction(title: "OK", style: .default))
+                            //                            self.present(ac, animated: true)
+                        }
+                    }
+                }
+            } else {
+                removeBlurView()
+                print("NO Biometrics")
+            }
+        } else {
+            removeBlurView()
+        }
+        print(Date().timeIntervalSince(startBack))
+    }
 }
 
 extension UIView {