+++ ReviewMonitor/ServiceCaller.swift
@@ -11,11 +11,11 @@ import UIKit
 class ServiceCaller: NSObject {
 
     private static let BaseUrl: String = {
-        /*#if DEBUG
-            return "http://127.0.0.1:4567/"
-        #else
-            return "https://review-monitor.herokuapp.com/"
-        #endif*/
+        /* #if DEBUG
+         return "http://127.0.0.1:4567/"
+         #else
+         return "https://review-monitor.herokuapp.com/"
+         #endif */
         "https://review-monitor.herokuapp.com/"
     }()
 
@@ -80,6 +80,7 @@ class ServiceCaller: NSObject {
         }
         url += "?" + convertToUrlParameter(params)
         url = url.addingPercentEncoding(withAllowedCharacters: CharacterSet.urlQueryAllowed)!
+        print(url)
         var request = URLRequest(url: URL(string: url)!)
         request.httpMethod = httpMethod.rawValue
         URLSession.shared.dataTask(with: request) { d, r, e in